requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
reportlab>=3.6.0
Pillow>=9.0.0
persiantools>=3.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.0
